feature: 2 - Product, flow create and get.

+++ src/Ambev.DeveloperEvaluation.Domain/Entities/Product.cs
@@ -12,7 +12,7 @@ public class Product : BaseEntity
     public string Description { get; set; } = string.Empty;
     public string Category { get; set; } = string.Empty;
     public string Image { get; set; } = string.Empty;
-    public int RatingRate { get; set; } = 0;
+    public decimal RatingRate { get; set; } = 0;
     public int RatingCount { get; set; } = 0;
 
     public ValidationResultDetail Validate()
